Correct the default kind of ids so that the code to implement the `hashCode()` method in generated classes uses the simpler integer variant.

+++ processor/src/test/resources/expected/com/example/collections/Arez_AbstractNonnullListObservableModel.java
@@ -174,7 +174,7 @@ public final class Arez_AbstractNonnullListObservableModel extends AbstractNonnu
   @Override
   public final int hashCode() {
     if ( Arez.areNativeComponentsEnabled() ) {
-      return Double.hashCode( $$arezi$$_id() );
+      return Integer.hashCode( $$arezi$$_id() );
     } else {
       return super.hashCode();
     }
